pyproject.toml
[project]
name = "uav_model"
version = "0.0.1"
description = "Repository to train segmentation model for UAV"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Alexandr G."}
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Programming Language :: Python"
]

dependencies = [
    "numpy>2.0",
    "opencv-python",
    "lightning",
    "torchvision",
    "segmentation-models-pytorch",
    "albumentations",
    "comet-ml",
    "tensorboard",
    "tensorboardX",
    "onnx"
]

[project.optional-dependencies]
dev = [
    "pip-tools",
    "pre-commit",
    "ipykernel",
    "matplotlib"
]

[project.urls]
dataset = "https://uavid.nl"